# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ShopMaster API
  version: 0.1.0
  description: |
    API REST de marketplace multi-lojista. Domínio modelado em DDD com
    bounded contexts; comprador, lojista e plataforma consomem os mesmos
    recursos, separados por **autorização**, não por formato de resposta.

    A autorização tem duas camadas: o papel de plataforma (`buyer`,
    `seller`, `platform_admin`) viaja no token; o papel de loja (`owner`,
    `admin`, `finance`) é resolvido por requisição, a partir do `storeId`
    da rota.

    Este arquivo é o **contrato**, mantido à mão — não é gerado por anotação
    no código. Rota que não está aqui não existe para quem consome a API.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  # Produção primeiro: é o que o Swagger UI seleciona por padrão, e é o
  # endereço que o time de front consome. O caminho `/shopmaster` faz parte
  # da URL base porque a VPS hospeda várias aplicações, cada uma num prefixo
  # — o Traefik o remove antes de encaminhar, e a app não o vê nas rotas.
  - url: https://jpdevm.systems/shopmaster
    description: Produção
  - url: http://localhost:8001
    description: Desenvolvimento local (docker compose up)
...
paths:
...
  /v1/me:
    get:
      tags: [Identity]
      operationId: getMyProfile
      summary: Minha conta
      description: |
        Devolve a conta do portador do token. O id vem **do token**, nunca da
        rota nem do corpo — não há como apontar para a conta de outra pessoa.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Token ausente, inválido ou expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: A conta do token não existe mais.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags: [Identity]
      operationId: updateMyProfile
      summary: Editar minha conta
      description: |
        Altera nome, telefone e documento (RF-005). **Parcial**: o que não for
        enviado não é tocado.

        **O e-mail não se troca no v1.** Enviá-lo devolve `422
        validation_failed` — recusar explicitamente é melhor do que ignorar em
        silêncio, senão quem tentou sai achando que trocou.

        **O documento só se define uma vez.** Depois de gravado, enviar um
        valor diferente devolve `409 document_already_set`. Reenviar o
        **mesmo** valor é aceito, para o PATCH continuar idempotente quando o
        cliente manda o formulário inteiro.

        `document` e `document_type` são obrigatórios um com o outro: um valor
        sem tipo é algo que ninguém sabe validar.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
            examples:
              contato:
                summary: Telefone
                value:
                  phone: (11) 98888-7777
              documento:
                summary: Definir o documento
                value:
                  document: 529.982.247-25
                  document_type: cpf
      responses:
        '200':
          description: Conta atualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Token ausente, inválido ou expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: A conta do token não existe mais.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: |
            `document_already_set` — a conta já tem documento, e ele é
            imutável. Reenviar o mesmo valor **não** cai aqui.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: document_already_set
                  message: O documento ja foi definido e nao pode ser trocado.
        '422':
          description: |
            | `error.code` | Quando |
            |---|---|
            | `validation_failed` | Campo mal formado, `document` sem `document_type`, tipo fora do enum, ou tentativa de trocar o e-mail. Traz `error.fields`. |
            | `document_already_in_use` | O documento já está cadastrado em outra conta. |
            | `invalid_document` | Formato ou dígito verificador não conferem. |
            | `invalid_phone` | DDD inexistente, tamanho errado, ou celular que não começa com 9. |
            | `invalid_name` | Nome vazio ou longo demais. |
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/ValidationError'
                  - $ref: '#/components/schemas/Error'
  /v1/me/addresses:
    get:
      tags: [Customer]
      operationId: listMyAddresses
      summary: Listar meus endereços
      description: |
        Agenda do portador do token (RF-006). O padrão vem primeiro.
      responses:
        '200':
          description: Lista (pode ser vazia).
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Address'
        '401':
          description: Token ausente, inválido ou expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [Customer]
      operationId: addMyAddress
      summary: Adicionar endereço
      description: |
        Inclui um endereço na agenda. O **primeiro** da conta vira padrão
        automaticamente. Pedir `is_default: true` num endereço seguinte
        troca o padrão na mesma gravação.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreAddressRequest'
            example:
              label: Casa
              recipient_name: Joao Pedro
              street: Av Paulista
              number: '1000'
              neighborhood: Bela Vista
              city: Sao Paulo
              state: SP
              postal_code: 01310-100
      responses:
        '201':
          description: Endereço criado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
        '401':
          description: Token ausente, inválido ou expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: |
            | `error.code` | Quando |
            |---|---|
            | `validation_failed` | Campo faltando ou UF fora do enum. Traz `error.fields`. |
            | `invalid_postal_code` | CEP sem 8 dígitos. |
            | `invalid_address` | Campo obrigatório vazio ou longo demais. |
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/ValidationError'
                  - $ref: '#/components/schemas/Error'
  /v1/me/addresses/{addressId}:
    parameters:
      - name: addressId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Customer]
      operationId: getMyAddress
      summary: Detalhe de um endereço
      description: |
        Endereço da própria agenda. Id de outra pessoa devolve o mesmo 404
        de inexistente — 403 confirmaria que o id é válido.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
        '401':
          description: Token ausente, inválido ou expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Não existe, ou não é do portador do token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [Customer]
      operationId: updateMyAddress
      summary: Editar endereço
      description: |
        Parcial. `is_default: true` troca o padrão. `label` ou `complement`
        com `null` remove o campo.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAddressRequest'
            example:
              is_default: true
      responses:
        '200':
          description: Endereço atualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
        '401':
          description: Token ausente, inválido ou expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Não existe, ou não é do portador do token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: |
            `validation_failed`, `invalid_postal_code` ou `invalid_address`.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/ValidationError'
                  - $ref: '#/components/schemas/Error'
    delete:
      tags: [Customer]
      operationId: removeMyAddress
      summary: Remover endereço
      description: |
        Se o removido era o padrão e ainda restam outros, o mais antigo
        restante vira o novo padrão — a agenda nunca fica sem padrão enquanto
        houver endereço.
      responses:
        '204':
          description: Removido.
        '401':
          description: Token ausente, inválido ou expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Não existe, ou não é do portador do token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token emitido por `POST /v1/auth`. Carrega os papéis de
        plataforma do usuário — **não** os papéis de loja, que são
        resolvidos por requisição.
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Código estável, em snake_case. É por ele que o cliente decide o que fazer.
              examples: [product_not_found]
            message:
              type: string
              description: Texto legível. Pode mudar sem aviso; não use como chave.
              examples: ['Produto 42 nao encontrado.']
...
    DocumentType:
      type: string
      enum: [cpf, cnh, cnpj]
      description: |
        O tipo decide **como** o valor é validado — e um CPF e uma CNH têm os
        dois 11 dígitos, então adivinhar pelo formato seria ambíguo.

        Os três têm dígito verificador conferido.

        O **CNPJ** segue o formato alfanumérico da IN RFB 2.229/2024: 12
        posições alfanuméricas de base + 2 dígitos verificadores, que
        continuam sempre numéricos. É superconjunto do formato antigo — todo
        CNPJ numérico já emitido continua válido pelo mesmo cálculo.

        Passaporte **não** é aceito, de propósito: não tem checksum, nem o
        brasileiro nem os estrangeiros. Aceitá-lo daria aparência de
        verificação a um campo que ninguém consegue verificar.
    Document:
      type: object
      required: [type, value]
      properties:
        type:
          $ref: '#/components/schemas/DocumentType'
        value:
          type: string
          description: Normalizado — só dígitos para CPF/CNH, maiúsculas sem máscara para CNPJ.
          examples: ['52998224725', 'AB12CD34EFGH83']
    BrazilianState:
      type: string
      enum:
        [AC, AL, AP, AM, BA, CE, DF, ES, GO, MA, MT, MS, MG, PA, PB, PR, PE, PI, RJ, RN, RS, RO, RR, SC, SP, SE, TO]
      description: UF brasileira.
    Address:
      type: object
      description: |
        Endereço da agenda do comprador. Não traz `user_id` — o dono é sempre
        o portador do token.
      required:
        [id, label, recipient_name, street, number, complement, neighborhood, city, state, postal_code, postal_code_formatted, is_default]
      properties:
        id:
          type: string
          format: uuid
        label:
          type: [string, 'null']
          examples: [Casa]
        recipient_name:
          type: string
        street:
          type: string
        number:
          type: string
        complement:
          type: [string, 'null']
        neighborhood:
          type: string
        city:
          type: string
        state:
          $ref: '#/components/schemas/BrazilianState'
        postal_code:
          type: string
          description: Só dígitos.
          examples: ['01310100']
        postal_code_formatted:
          type: string
          examples: ['01310-100']
        is_default:
          type: boolean
    StoreAddressRequest:
      type: object
      required: [recipient_name, street, number, neighborhood, city, state, postal_code]
      properties:
        label:
          type: [string, 'null']
          maxLength: 40
        recipient_name:
          type: string
          maxLength: 120
        street:
          type: string
          maxLength: 200
        number:
          type: string
          maxLength: 20
        complement:
          type: [string, 'null']
          maxLength: 120
        neighborhood:
          type: string
          maxLength: 120
        city:
          type: string
          maxLength: 120
        state:
          $ref: '#/components/schemas/BrazilianState'
        postal_code:
          type: string
          description: Aceita máscara; grava só dígitos.
        is_default:
          type: boolean
          description: Se true, este passa a ser o padrão (desmarca o anterior).
    UpdateAddressRequest:
      type: object
      description: Parcial — o que não for enviado não é alterado.
      properties:
        label:
          type: [string, 'null']
        recipient_name:
          type: string
        street:
          type: string
        number:
          type: string
        complement:
          type: [string, 'null']
        neighborhood:
          type: string
        city:
          type: string
        state:
          $ref: '#/components/schemas/BrazilianState'
        postal_code:
          type: string
        is_default:
          type: boolean
...
    UserProfile:
      type: object
      description: |
        A conta como o **dono** dela a vê. O documento aparece por inteiro
        porque quem pede é o titular. Nunca reaproveite este formato em rota
        de terceiro — CPF é dado sensível.

        **Não traz `roles`.** Papel é autorização, não perfil: o banco pode já
        ter um papel que o token ainda não tem, e devolvê-lo aqui prometeria
        uma permissão que o middleware vai recusar. Para saber o que o
        portador pode fazer **agora**, use `GET /v1/auth/session`.
      required: [id, name, email, phone, phone_formatted, document]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: [string, 'null']
          description: Só dígitos, DDD incluído.
          examples: ['11988887777']
        phone_formatted:
          type: [string, 'null']
          examples: ['(11) 98888-7777']
        document:
          oneOf:
            - $ref: '#/components/schemas/Document'
            - type: 'null'
    UpdateProfileRequest:
      type: object
      description: |
        Parcial — o que não for enviado não é alterado. Corpo vazio é aceito e
        devolve o perfil inalterado, para o cliente que reenvia o formulário
        sem mudanças não precisar tratar um erro.
      properties:
        name:
          type: string
          maxLength: 120
        phone:
          type: [string, 'null']
          description: |
            Aceita máscara e `+55`; é normalizado para só dígitos.

            Três estados: **omitir** não mexe, **um valor** substitui, e
            **`null` explícito remove** o telefone da conta (RFC 7396).
          examples: ['(11) 98888-7777']
        document:
          type: string
          description: Obrigatório junto de `document_type`. Imutável depois de definido.
          examples: ['529.982.247-25']
        document_type:
          $ref: '#/components/schemas/DocumentType'
...
    ValidationError:
      type: object
      description: |
        Mesma casca do `Error`, com o detalhe por campo. `fields` é acréscimo:
        um cliente que só conheça `code` e `message` continua funcionando.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, fields]
          properties:
            code:
              type: string
              enum: [validation_failed]
            message:
              type: string
            fields:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
...
security:
  - bearerAuth: []